test(#450): migrate PreConditionForm tests to TypeScript
Migrate all 56 tests from .test.jsx to .test.tsx with async
adaptations for the RHF + Zod migration:

- Wrap field-change assertions (threshold, cooldown, sensor_type,
  comparison, time inputs) in waitFor() for RHF async validation
- Use last-call pattern for onChange assertions since RHF may
  fire multiple propagation cycles
- Add 4 new parent-error-wiring tests (threshold, cooldown,
  sensor_type, comparison aria-invalid/aria-describedby)
- Add prop-sync test verifying form updates on external prop change
- Fix toggle-off-and-on error-clearing test to properly simulate
  parent state transitions (null -> default)
- Type all PreConditionValue literals and mock functions
- Use props() factory pattern consistent with TimeWindowInput tests
